requirements.txt
web3==7.6.0
python-dotenv
numpy
websocket-client
